Add CondaEcosystemModifyDllSearchPath()
This changes the DLL search order so that C:\Windows\System32 does not
get searched in before entries in PATH.

Modified twice and backported.

Now add CWD and WinDir.

I observed that you must call LoadLibraryExW() with the fullpath of
the DLL for this stuff to not fail. It is probably worthwhile checking
in procmon what directories get looked in in the non-fullpath case.